requirements.txt
fastapi
uvicorn
pymilvus
openai
python-dotenv
tiktoken
PyMuPDF
python-docx
streamlit
requests
